backend/requirements.txt
fastapi
uvicorn
httpx
Pillow
numpy